Sad attempt at better Zebra queue
The Zebra library's request queue turns out to not be great, because it only sends one request at a time, and waits for the request to finish before sending the next one. As a result, the printer pauses momentarily between every barcode format command.

feed_printer() is a sorry attempt at keeping the printer's receive buffer full, without potentially overfilling it, and in a way that the items in the barcode_queue are removed roughly when they're being printed.