Basic multi-module functionality
Allows user to open a module browser in a new tab, rather than navigating to a different /#/module/address and closing all tabs that are currently open.

A few changes still need to happen, such as adding an 'x' to Record Browse tabs, giving the user to close a browse tab. Currently that's not an option because RW expects only one module browse tab to be available at a time.